Move molgenis client code to molgenis package
