+++ mcmd/commands/make.py
@@ -1,6 +1,6 @@
 from mcmd import io
-from mcmd.client import api
-from mcmd.client.molgenis_client import post, get
+from mcmd.molgenis import api
+from mcmd.molgenis.client import post, get
 from mcmd.command import command
 from mcmd.commands._registry import arguments
 from mcmd.io import highlight
