+++ mcmd/molgenis/request_handler.py
@@ -6,7 +6,7 @@ Error responses can come back in varying forms which this decorator tries to uni
 
 import requests
 
-from mcmd.client import auth
+from mcmd.molgenis import auth
 from mcmd.utils.errors import McmdError, MolgenisOfflineError
