+++ tests/unit/version/molgenis_version_test.py
@@ -3,7 +3,7 @@ import unittest
 import pytest
 
 from mcmd.utils.errors import McmdError
-from mcmd.version import molgenis_version
+from mcmd.molgenis import version
 
 
 @pytest.mark.unit
@@ -11,24 +11,24 @@ class MolgenisVersionTest(unittest.TestCase):
 
     @staticmethod
     def test_extract_version_number():
-        version = molgenis_version._extract_version_number('7.0.0')
+        version = version._extract_version_number('7.0.0')
         assert version == '7.0.0'
 
     @staticmethod
     def test_extract_version_number_snapshot():
-        version = molgenis_version._extract_version_number('8.0.1-SNAPSHOT')
+        version = version._extract_version_number('8.0.1-SNAPSHOT')
         assert version == '8.0.1'
 
     @staticmethod
     def test_extract_version_number_preview():
-        version = molgenis_version._extract_version_number('1.2.3-SNAPSHOT-PR-1234-2')
+        version = version._extract_version_number('1.2.3-SNAPSHOT-PR-1234-2')
         assert version == '1.2.3'
 
     @staticmethod
     def test_extract_version_number_multi():
-        version = molgenis_version._extract_version_number('7.2.3-APP-1.0.0')
+        version = version._extract_version_number('7.2.3-APP-1.0.0')
         assert version == '7.2.3'
 
     def test_extract_version_number_invalid(self):
         with self.assertRaises(McmdError):
-            molgenis_version._extract_version_number('4.0-TESTING')
+            version._extract_version_number('4.0-TESTING')
